refactor(x/ecocredit): migrate ecocredit utils to core (#1048)
* refactor(x/ecocredit) migrate ecocredit utils to core

* refactor(x/ecocredit) migrate ecocredit utils to core

* refactor(x/ecocredit) migrate ecocredit utils to core

* use core for query batch
