+++ x/ecocredit/basket/msg_take.go
@@ -6,6 +6,7 @@ import (
 	"github.com/cosmos/cosmos-sdk/x/auth/legacy/legacytx"
 
 	"github.com/regen-network/regen-ledger/x/ecocredit"
+	"github.com/regen-network/regen-ledger/x/ecocredit/core"
 )
 
 var _ legacytx.LegacyMsg = &MsgTake{}
@@ -37,7 +38,7 @@ func (m MsgTake) ValidateBasic() error {
 		return sdkerrors.ErrInvalidRequest.Wrapf("%s must be positive", m.Amount)
 	}
 	if m.RetireOnTake {
-		if err := ecocredit.ValidateJurisdiction(m.RetirementJurisdiction); err != nil {
+		if err := core.ValidateJurisdiction(m.RetirementJurisdiction); err != nil {
 			return sdkerrors.ErrInvalidRequest.Wrap(err.Error())
 		}
 	}
