+++ x/ecocredit/core/params.go
@@ -3,8 +3,6 @@ package core
 import (
 	"regexp"
 
-	"github.com/regen-network/regen-ledger/x/ecocredit"
-
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
 	paramtypes "github.com/cosmos/cosmos-sdk/x/params/types"
@@ -117,7 +115,7 @@ func validateCreditTypes(i interface{}) error {
 	seenAbbrs := make(map[string]bool)
 	for _, creditType := range creditTypes {
 		// Validate name
-		T := ecocredit.NormalizeCreditTypeName(creditType.Name)
+		T := NormalizeCreditTypeName(creditType.Name)
 		if T != creditType.Name {
 			return sdkerrors.ErrInvalidRequest.Wrapf("credit type name should be normalized: got %s, should be %s", creditType.Name, T)
 		}
