+++ x/ecocredit/server/core/query_batch_info.go
@@ -6,13 +6,12 @@ import (
 	sdk "github.com/cosmos/cosmos-sdk/types"
 
 	"github.com/regen-network/regen-ledger/types"
-	"github.com/regen-network/regen-ledger/x/ecocredit"
 	"github.com/regen-network/regen-ledger/x/ecocredit/core"
 )
 
 // BatchInfo queries for information on a credit batch.
 func (k Keeper) BatchInfo(ctx context.Context, request *core.QueryBatchInfoRequest) (*core.QueryBatchInfoResponse, error) {
-	if err := ecocredit.ValidateDenom(request.BatchDenom); err != nil {
+	if err := core.ValidateDenom(request.BatchDenom); err != nil {
 		return nil, err
 	}
